autotest: Account for ARP requests in arping result parsing
Under some circumstances, we see requests coming back from devices.
This adds a slightly different twist on the output format that we'll
need to account for because we check that we saw all the responses that
arping thinks it saw.

TEST=network_WiFi_GTK still passes.  This is a low percentage failure.
We should just stop seeing sporadic GTK test failures.
BUG=None

Change-Id: I160d03018ce38dfcab6ede28ef4dcf869e713203
Reviewed-on: https://gerrit.chromium.org/gerrit/61423
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>