VSWR: support reading parameters directly from shopfloor
Adding the feature to directly loading parameters from shopfloor. With
this feature, we should be more easily to control parameters. We still
keep the USB appraoch as a backup both in the sense of loading
configuration and storing logs.

BUG=chrome-os-partner:20276
TEST=Manually create a test_list with load_from_shopfloor dargs added.
Test both shopfloor and USB parameters are loaded correctly.

Change-Id: Ibbc8d9566d1a28b29e56c214c2732f3f03244832
Reviewed-on: https://gerrit.chromium.org/gerrit/61569
Reviewed-by: Jon Salz <jsalz@chromium.org>
Tested-by: Chun-Ta Lin <itspeter@google.com>
Commit-Queue: Chun-Ta Lin <itspeter@google.com>