Add a results database migration to create the tko_iteration_perf_value table.
This change is part of a series of CLs in which we're changing how
autotest stores and handles perf test data.  Previously, we added a
new function for perf tests to output perf measurements to an output
file.  In the current CL, we're adding a new results database table.
In a subsequent CL, this table will be used by the TKO parser to store
the outputted perf measurements.

BUG=chromium:258229
DEPLOY=migrate (/usr/local/autotest/database$ python migrate.py sync 78)
TEST="python migrate.py sync 78" to verify using a local autotest database
         that the new table gets created.
     "python migrate.py sync 77" to verify that the new table is dropped if
         we migrate back down.

Change-Id: I210a42b3c3c5ec1b7aa26dd0caabda5f6e2e2bb9
Reviewed-on: https://gerrit.chromium.org/gerrit/61706
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>