Autotest: Fix reimager successful or not logic.
Currently if the reimaging job times out, its not checking whether
or not enough machines finished reimaging to continue with the suite.

Fixed the logic flow so this check now occurs successfully.

BUG=chromium:259878
TEST=local setup. Queued up a 3 machine tryjob with 1 machine
guaranteed to stay stuck in Queued and not run. Ensured the fix
continues the suite once the timeout is hit.

Change-Id: I0063e071e743ac46bdaf9e8f19cf7e2892687d1a
Reviewed-on: https://gerrit.chromium.org/gerrit/61698
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>