autotest: Test shill behavior in client tests
We have a number of tests in network_WiFiManager that make assertions
about what types of SSIDs/passwords shill will accept as valid, and what
kinds of errors shill with throw on invalid parameters.  Move these to
the more appropriate client test because we don't actually need a router
to make this happen.

TEST=This adds two passing tests.
BUG=chromium:259487
Change-Id: Ifab36af33f34e905340509d59e0143744eb234ef
Reviewed-on: https://gerrit.chromium.org/gerrit/62136
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>