Fix unit test to actually run against an AP.
The AP has to be powered on via the configurator object.  If it is not
the configurator will short circut and not really do anything.

As for the test_invalid_security this will fix that is will pass for
most APs.  Some APs actually change the mode automatically if you do
WEP, so they don't throw an error.  Since this test is run by the
creator of the configurator I think it is fine to handle this on a
case by case basis.

TEST=Ran it
BUG=chromium:232220

Change-Id: I19113f7aa64ef2e1feb8c35180a5b251ccd69f14
Reviewed-on: https://gerrit.chromium.org/gerrit/62031
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Tan Gao <tgao@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>