firmware_TouchMTB: print raw metrics values
In the last patch, the statistics of metrics values are printed
as follows:

Metrics statistics by validator  :    fw_11.23  description
--------------------------------------------------------------------
  Linearity(BothEnds)Validator
    max error in x (mm)        :         9.34   The max err of all samples
    ...

If a user would like to know which gesture/file causes the max error
above, it would be helpful to print the raw metrics values as follows:

Raw metrics values
--------------------------------------------------------------------
fw_11.23

  Linearity(BothEnds)Validator
    max error in x (mm)
      drag_edge_thumb.left_to_right (20130710_063117)          : 3.06
      two_fat_fingers_tracking.left_to_right (20130710_063117) : 9.34
      ...

In this way, the user could observe that the file with
  gesture: "two_fat_fingers_tracking",
  variation: "left_to_right",
  in the folder: "20130710_063117"
causes the "max error in x".

BUG=chromium:235753
TEST=Run the command on a chromebook machine to generate a summary report.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Replay the logs
$ tools/machine_replay.sh -b tests/logs/lumpy -s

Generate the summary report with short metrics values grouped by validator.
$ ./summary.sh -d /tmp -m

Generate the summary report with detailed metrics values.
$ ./summary.sh -d /tmp --metrics_details
Observe that the raw detailed metrics values are printed this time.

Change-Id: I4d540f71c1c1e91507b9104eff8ac80c14d6654e
Reviewed-on: https://gerrit.chromium.org/gerrit/62329
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>