Autotest TKO parser now parses perf_measurements test output files.
This change is part of a series of changes in which we're modifying how
autotest handles/stores perf test measurements, to better align with chrome
team's perf dashboard.

Previously, a new function was added that tests can invoke to output
perf measurements; this information is written into a test result file
named "perf_measurements".

The next step is to modify the autotest TKO parser to parse the
perf_measurements output file and then insert this data into the results
database.  This first half of this -- having the TKO parser parse the
perf_measurements output file -- is accomplished in this CL.  In the
next CL, the parsed information will be inserted into the results
database.

Changes in the current CL are modeled after the way in which Autotest
currently parses performance keyvals.

BUG=chromium:258230
TEST=(1) Verified all unit tests passed: "python version_1_unittest.py".
     (2) Invoked platform_GesturesRegressionTest using a local AFE and
         verified that the results in the "perf_measurements" output file
         are parsed correctly.
     (3) Invoked the parsing code on a manually-constructed "perf_measurements"
         output file containing results from multiple iterations, and a mixture
         of individual values and lists of values, and verified that the perf
         data is parsed correctly.
     (4) Invoked platform_BootPerf using a local AFE (which currently does
         not result in a "perf_measurements" output file) and verified that
         the TKO parser still passes successfully.

Change-Id: I88090c980cea0c8e09dc0156911dcc357922c27a
Reviewed-on: https://gerrit.chromium.org/gerrit/62672
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>