[autotest] Email notifications when job_repo_url changes on a host.
Everytime a devserver flakes while verifying job_repo_url we end
up re-staging contents. This leads to a delay in the suite run
time. To make it clean when such a delay is incurred, this cl
sends out an email everytime the job_repo_url changes during
verification.

Previously this information was logged, but checking that it
occured involved going through the logs of each test in a
suite which is tedious for suites like bvt.

TEST=Set job_repo_url and verified that an email gets sent
     out with the appropriate message when it changes.
BUG=chromium:261258

Change-Id: I7f64f63e908b3bda60896e3b6944f6db346a78a1
Reviewed-on: https://gerrit.chromium.org/gerrit/62557
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>