Make background window last entire power_LoadTest
The comment for the background window says that the window should last for the
entire session, but it only lasts for 2 mintes. This fixes that and makes sure
that the background window doesn't grab the foreground in the UI like it was.
Make sure that the focus is not on the background window.

BUG=chromium:261943
TEST=power_LoadTest
	make sure the background window is not in the UI foreground for longer than
a second. Also check that the window lasts for longer than 2 minutes.

Change-Id: I6c77d3ef2fa1697cb2c2166cd6187c678969d7cd
Reviewed-on: https://gerrit.chromium.org/gerrit/62588
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>