[autotest] add a --no-quickmerge flag to test_that
This CL adds a --no-quickmerge flag to test_that, which will force it to
skip the quickmerge step. This option will be used for test_that runs on
the builder (as part of VMtest), since on the builder we will have a
fresh sysroot and there is no need to run quickmerge.

BUG=chromium:262135
TEST=Unit tests pass. Manually verified that flag works.

Change-Id: Ia1e12257f5c1288e352f0d3e0078ed764688b0fd
Reviewed-on: https://gerrit.chromium.org/gerrit/62648
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>