[autotest] add a --no-experimental flag to test_that
This CL adds a --no-experimental flag to test_that, which will cause it
to skip tests marked as experimental when scheduling a suite. A minor
change to suite.py was required in order for test_that's count of
scheduled tests to be correct when skipping experimental tests.

BUG=chromium:262135
TEST=Manually verified that experimental tests do not get scheduled when
flag is used. Unit tests pass.

Change-Id: I33247446e95915f372ced59f46b78c36fa82c1a7
Reviewed-on: https://gerrit.chromium.org/gerrit/62657
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>