Destroy the webdriver connection at all times.
Previously if configuration failed, the exception was thrown caught
but the webdriver connection was maintained.  That means the chrome
window was still open.  Now we will attempt to always destroy the
connection which should close the window.

In addition, if the only command is to power down the AP, don't
connect to webdriver and use the rpm infrastructure to power down
the AP.

TEST=Manual, ran it
BUG=None

Change-Id: I76926a109e43087576cffe03cb525ba8a7e1d8ad
Reviewed-on: https://gerrit.chromium.org/gerrit/62705
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Viral Modi <viralmodi@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>