Always open new Chrome window at end of loop
There's a race condition where we check if there are any Chrome windows open
then uninstall the power_LoadTest extension. The race happens when there is a
window open when we check if there are any windows open, but it closes before we
uninstall the extension. The exception produced by this causes the test to fail
and prevent the results from being recorded.

BUG=chrome-os-partner:18324
TEST=Put sleep(30) between opening window and uninstalling extension. Make sure
exception is not thrown. When running power_LoadTest

Change-Id: Icf35a5adba7f088aa32685bc4f3c0d4935a27ac6
Reviewed-on: https://gerrit.chromium.org/gerrit/63146
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>