Bluetooth: add basic adapter functionality to client
Just about all tests will want to reset the device to a known state
so provide reset_on() and reset_off() to do that. The difference is
whether the adapter is powered on or off after the reset; most tests
will want the former, but it's conceivable the latter might be
necessary.

Provide basic property manipulation with set_powered(),
set_discoverable() and set_pairable().

BUG=chromium:256771
TEST=simple test that uses the exposed methods,
 verify operation using btmon on DUT

Change-Id: Ib3a5efb3ea5de8a457e9043f57801d7b00cf1808
Reviewed-on: https://gerrit.chromium.org/gerrit/63261
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>