Do not raise an ImportError when attempting to translate a subprocess exception.
After running a client-side test, autotest attempts to parse and re-raise any
exceptions raised by the test subprocess.  This is done by attempting to
un-pickle a text file that may contain information about the raised exception.
However, the un-pickling may fail in the event that the subprocess raises a
subprocess-specific exception that autotest proper doesn't know about
(e.g., pyauto raising a pyauto_errors.JSONInterfaceError).  This causes the
immediate autotest failure to be an "ImportError" that actually masks the
root-cause failure that was originally raised by the subprocess itself.

In this CL, this situation is captured by catching the ImportError exception,
then re-raising a general error.TestError with an informative error message
indicating that the root cause exception occurred elsewhere (it also prints
out the text contents of the file that was attempted to be un-pickled, which
isn't completely human-readable, but does contain text about the root cause
exception that was originally raised.  Here's an example failure message for
this case:

10:05:48 INFO | 		ERROR	login_CryptohomeUnmounted	login_CryptohomeUnmounted	timestamp=1374771948	localtime=Jul 25 10:05:48	Subprocess raised an exception that could not be identified. The root cause exception is in the text that follows: cautotest_lib.client.common_lib.error
  TestFailRetry
  p0
  (cpyauto_errors
  JSONInterfaceError
  p1
  (S'Raised exception message here.'
  p2
  tp3
  Rp4
  tp5
  Rp6
  .
10:05:48 INFO | 	END ERROR	login_CryptohomeUnmounted	login_CryptohomeUnmounted	timestamp=1374771948	localtime=Jul 25 10:05:48

BUG=chromium:244593
TEST=Ran login_CryptohomeUnmounted on a local device in which I forced the
underlying pyauto code to raise a pyauto_errors.JSONInterfaceError, then
verified that the example error log above was the result.

Change-Id: I3132f89dd962e2f266b65a8d96fc87ac05075649
Reviewed-on: https://gerrit.chromium.org/gerrit/63355
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>