firmware_TouchMTB: add the metric "clicks without finger IDs"
In this patch, we add a new metric to PhysicalClickValidator
"of the n clicks, the % of clicks without a finger ID".
In particular, we check whether the BTN_LEFT event occurs when
the number of tracking (finger) IDs is correct. For example,
if the BTN_LEFT events with values 1 and 0 occur after a finger
has left, it is a failure. The event sequence in such a failure
case may look like

  Event: ABS_MT_TRACKING_ID, value 28
  Event: ABS_MT_TRACKING_ID, value -1
  Event: BTN_LEFT, value 1
  Event: BTN_LEFT, value 0

On the other hand, a click case with the correct number of tracking
IDs may look like

  Event: ABS_MT_TRACKING_ID, value 28
  Event: BTN_LEFT, value 1
  Event: ABS_MT_TRACKING_ID, value -1
  Event: BTN_LEFT, value 0

BUG=chromium:251193
TEST=Replay logs on a lumpy.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ tools/machine_replay.sh -b tests/logs/lumpy/ -s

And then run the summary script to generate a summary report.
$ ./summary.sh -d /tmp -m f
Observe the statistics of the metrics would look like
PhysicalClickValidator
  1f-click miss rate (%)     : 12.50  the pct of finger IDs w/o a click
  1f-click w/o finger IDs (%):  0.00  the pct of clicks w/o correct finger IDs
  2f-click miss rate (%)     : 75.00  the pct of finger IDs w/o a click
  2f-click w/o finger IDs (%):  0.00  the pct of clicks w/o correct finger IDs

The raw metrics values look like
PhysicalClickValidator
  1f-click miss rate (%)
    ** Note: value below represents (actual clicks, expected clicks)
    one_finger_physical_click.bottom_side (20130506_032458)  : (1, 1)
    ...
  1f-click w/o finger IDs (%)
    ** Note: value below represents
             (clicks with correct finger IDs, actual clicks)
    one_finger_physical_click.bottom_left (20130506_032458)  : (1, 1)
    ...

Also all of the unit tests should pass.
$ python tests/run_all_unittests.py

Change-Id: I26b7e7dd6da4456f7b2fae8da7aacdfb5e69b688
Reviewed-on: https://gerrit.chromium.org/gerrit/63588
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>