[autotest] Make all suites server-side.
We hardcode suites to be server-side tests in the create_suite_job RPC,
but it appears a few cases of people setting their suite as a
client-side test has happened.  Just to make sure there's no confusion
as to what the right setting is, and so that nothing crazy happens in
the future, let's get these suites back to server-side as they should
be.

BUG=None
TEST=None

Change-Id: I9dc2cd01335117da99be7f03572f12f841998bf5
Reviewed-on: https://gerrit.chromium.org/gerrit/63509
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>