[autotest] Don't swap devservers when verifying repo_url.
Since we're seeing more flake than we would like when
checking the health of a devserver we end up swapping
devservers and restaging contents, even when the initial
devserver is healthy. This leads to more load, which
further exacerbates the problem of flakyness, leading to
a vicious cycle. This cl restages autotest on the
job_repo_url devserver, and fails verification if the
devserver embedded in the url is down.

BUG=None
TEST=set job_repo_url and deleted directories on local
     devserver, checked that we re-stage. Didn't start
     up local devserver and checked that we received
     an URLError. Killed the devserver in the middle
     of a request and checked that the test Failed.

Change-Id: I18843d3047122b9ffea3b2b1305fb0332172b19d
Reviewed-on: https://gerrit.chromium.org/gerrit/63758
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>