autotest: Rework netperf based test to adjust for measurement variance
Add a NetperfSession object which runs netperf for a fix number of
iterations or until the variance is observed to be below a threshold.
Consider a test to have passed if the assertion falls no more than one
standard deviation below the averaged result.

TEST=Observe test output fly by.
BUG=chromium:266166

Change-Id: I1040d15965ef83bbcdd44968bb2d7a264f14d13a
Reviewed-on: https://gerrit.chromium.org/gerrit/64379
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>