firmware_TouchMTB: fix the replay problem due to non-existence of raw data files
When in replaying mode, the program will replay existing data files
in the specified folder. If there is no existing data file(s) for a
gesture, the gesture will be skipped. Currently, there is a bug.
If a gesture is skipped owing to the lack of raw data files, the
gesture test results are still displayed using the previous mtplot
image without any validators test results in the html report.

In this patch, we fix this bug by resetting the gesture image
when all contents of current test has been flushed to the html
report. In this way, the gesture image will not be used repeatedly
by mistake.

BUG=chromium:266721
TEST=Replay the raw data files on a lumpy chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ tools/machine_replay.sh -b tests/logs/lumpy/ -s
The replay is expected to finish within half a minute. Check the
/tmp/touch_firmware_repot_xxxx.html report files using chrome.
It should display the html report correctly without any empty
validators test results and wrong mtplot images.

Change-Id: I281fc8c1ecdaf3c2a0fd8dd43de45ca59946c4c8
Reviewed-on: https://gerrit.chromium.org/gerrit/64179
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>