firmware_TouchMTB: reverse coordinates for touchscreen in robot mode
When performing the touchscreen tests using the robot, the device is
placed upside down and is rotated 180 degrees. We need to reverse the
coordinates to make correct gestures.

BUG=chromium:266720
TEST=conduct the touchscreen firmware tests with the touch robot.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Conduct the touchscreen tests with 1 iteration using the robot.
$ DISPLAY=:0 python main.py -m robot -i 1 --touchscreen
Observe that the robot performs the gestures correctly with respect
to the touchscreen's orientation.

Change-Id: Ic01f2612942585bb1f72da42ff986b8adb50306d
Reviewed-on: https://gerrit.chromium.org/gerrit/64180
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>