firmware_TouchMTB: use regular command line options
We used to use pyauto to display the html result file after finishing
a round of tests. Due to some side effect of pyauto, we need to keep
the command line options in an environment variable called OPTIONS.

The pyauto has now been replaced by telemetry which does not expose
the side effect. We could now use the more common and natural
command line options.

In this patch, we first check if there are command line options.
If no command line options could be found, we try to use the
environment variable OPTIONS that we used to use before. If no
such OPTIONS environment variable could be found either, we use
the default values.

BUG=chromium:266724
TEST=conduct the touch firmware tests by ssh-ing to a chromebook
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Conduct a complete round of tests with 3 iterations.
$ DISPLAY=:0 python main.py -m complete -i 3
Observe that it could launch the tests as before.
Press 'x' to terminate the tests anytime.

Note that the above command is equivalent to the old one below
which will be supported for a longer while:
$ DISPLAY=:0 OPTIONS="-m complete -i 3" python main.py

Change-Id: I5e21c7b76fc163e5cc69d180826f34f87380759f
Reviewed-on: https://gerrit.chromium.org/gerrit/64178
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>