firmware_TouchMTB: fix exceptions about the Report Rate Validtor
In some two finger gestures, e.g., with the first finger resting
on the pad, the 2nd finger draws a line, if somehow a tester
perform a only-one-finger gesture accidentally, the 2nd finger
would not exist. To calculate the report rate metrics for the
2nd finger would result in a division-by-0 error.

Another problem with the report rate validator is that if the
validator is specified as

  ReportRateValidator(report_rate_criteria, finger=0),

The current code would treat it as finger=None. We should only
extract the packets for finger=0 only. Note that the finger
contact number starts from 0.

BUG=chromium:267371
TEST=Conduct a manual test on a chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ DISPLAY=:0 python main.py -s --skip_html

When it prompts to perform a gesture for "finger_crossing", use only one
finger to perform a wrong gesture, e.g., draw a line. Observe that
- it does not incur a division-by-0 error,
- it shows a warning message in red on the window about the missing finger, and
- the score of the Report Rate Validator on the right frame of the window is 0

Change-Id: I749c4f0a77f4203ca517d2e486fd196d1817aa3b
Reviewed-on: https://gerrit.chromium.org/gerrit/64316
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>