Use proper anchor strings to time ARM platforms AP boot
The EC console output on pit does not include the ' PB ' lines this
autotest uses to time the beginning of the AP boot. Incidentally, the
'AP running' line is printed not when the AP is actually showing signs
of life, but when the EC finishes sequencing power.

As a temp fix, for ARM platforms use 'AP running' as indication of the
beginning of AP boot, and 'XPSHOLD seen' as indication of AP actually
executing instructions. There should be a better way of timing these
events, the same on all platforms.

BUG=chrome-os-partner:21628
TEST=manual
  $ run_remote_tests.sh --board=peach_pit --remote=192.168.1.11 --use_emerged \
        --servo firmware_ECBootTime/control

   now passes successfully.

  also observed that the anchor lines are generated on spring and snow
  ec console in response to 'power on' command.

Change-Id: I010c8c9d9e0faf315f9e93c4ae1f6af5d20bf6f1
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64723
Reviewed-by: Randall Spangler <rspangler@chromium.org>