autotest: Clean up default SSID logic
This causes stumpy cells to stop having SSIDs that start with
chromeos1-shelf...  This was happening because the path to the
ip command is different on stumpy cells, and this logic was hardcoding
this command.  Fix this by moving the building up of default SSIDs to
inside the Linux router where it rightly belongs.

BUG=None
TEST=Ran wifi_matfunc suite, which shows that both old and new style
tests are compatible with this change.  Ran a test out of
network_WiFiPerf which actually triggers the SSID generation logic
to prove that older WiFi tests also benefit from this change.

Change-Id: Ib33407503a0a1c6d8922d84e9c4458be43a6dd06
Reviewed-on: https://gerrit.chromium.org/gerrit/64955
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>