security_x86Registers: validate Atom registers
The Atom processors were failing because they had no tests at all for
MSRs. Instead of that, check for the fact that they are missing those
MSRs entirely.

In addition, the GCS.BILD flag is unset on Atom devices, so a bug was
opened, and an exception added for now.

This change also adds checks to catch missing bit tests, and cleans up
the logging call syntax as pointed out by pylint.

BUG=chromium:269633
TEST=security_x86Registers passes again on Alex

Change-Id: Ib49ff06b689cb065730237ab4a7f0c90a0d8e6ee
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64964
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>