firmware_TouchMTB: adjustment of linearity validator
We need to do some adjustment for the linearity validator according
to the x-t and the y-t figure in the attached png file in
chromium:248210 #6 which is a left-to-right one-finger-tracking
gesture performed by the robot.

Adjustment as follows:
  (1) Calculate the regression line based on only the middle segment.
  (2) Calculate the metrics of max errors and rms erros only on the
      middle segment. The begin segment and the end segment may only
      represent a pause.
  (3) Pick up only certain axis(es) when calculating the statistics
      of metrics:
      horizontal lines: metrics only in the x axis
      vertical   lines: metrics only in the y axis
      diagonal   lines: metrics both in the x and y axes

BUG=chromium:248120
TEST=Run the unit tests on a chromebook or in chroot
(cr) $ cd .../firmware_TouchMTB
(cr) $ python tests/run_all_unittests.py
Observe that all unit tests should pass.

Change-Id: I28566662831aae50be3274bd9b23e870506ba160
Reviewed-on: https://gerrit.chromium.org/gerrit/66344
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>