firmware_TouchMTB: fix the replay error about max recursive depth exceeded
If we record the gestures using the robot with 3 iterations, and then
replay the gestures, it may result in the max recursive depth exceeded.

The gesture recording program is an event driven program in which the
X window system will receive and handle incoming events. However, our
replay program used a recursive function call to replay gesture files.
When the number of gesture files is large, this may result in an error
of max recursive depth exceeded.

A simple solution is to replay the gesture files in the same event
driven manner that is used when recording the gesture files. This
would reuse most of the flow control logic. This is achieved
primarily by registering a timeout callback in the method
_setup_this_gesture_variation() when replaying.

BUG=chromium:271279
TEST=Record gesture files with 3 iterations using the robot.
And then replay the gesture files on the same chromebook machine.
Observe that the replay finish correctly without the max recursive
depth exceeded problem.

Change-Id: Ib6cb2872d2e62b7e533bcf7e214dce57806e71bc
Reviewed-on: https://gerrit.chromium.org/gerrit/65560
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>