firmware_TouchMTB: use finger contact to replace slot for LinearityValidator
Previously, we assume that the first finger contact is always
assigned the slot 0, and the 2nd finger contact the slot 1, etc.
While this may hold true in practice, it is not safe to make the
assumption.

Hence, we would like to use the finger contacts to replace the slot numbers in
validators. For example,

  LinearityValidator(linearity_criteria, slot=0, segments=MIDDLE),

would be replaced by

  LinearityValidator(linearity_criteria, finger=0, segments=MIDDLE),

where finger means finger contact.

The difference between the slot number and the finger contact lies in
that the former is assigned by the firmware while the latter is how we
observe in the events emitted by the event device (/dev/input/event*).

BUG=chromium:271282
TEST=Conduct the unit tests on a chormebook or in chroot, they should
all pass.
$ cd .../firmware_TouchMTB
$ python tests/run_all_unittests.py

Change-Id: I4f4a6faeb0cf03cb404382f543dea52f03c2f715
Reviewed-on: https://gerrit.chromium.org/gerrit/65562
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>