[autotest] Fail, not abort, tests that fail to provision.
If we abort the test, something across dynamic_suites and run_suite will
refuse to look at the reason field to see if one is listed, and thus we
get no reason displayed on the run_suite output, and by proxy, the
waterfall.

Testing also revealed that another of my previous assertions, that we
need to set the complete bit on the HQE, was actually wrong, and will
also be done by the FinalReparseTask, so that code and comments are
being dropped.

BUG=chromium:188217
DEPLOY=scheduler
TEST=run_suite now shows useful output

Change-Id: I7369d3847687cd0f0daad50a4e5658eef96f39e4
Reviewed-on: https://gerrit.chromium.org/gerrit/65116
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>